Added support for the FluentValidation.NET library.
Dependency injection for validator constructors happens in the new
UnityValidatorFactory class in the Impl lib. The calling assembly is
scanned for all validators, and each is registered automatically, so that
no changes to the unity config are needed.

The factory is used in the fluent validation model validator provider
configuration in Global.asax. Previous DataAnnotations validators still
work.